RIP to Float16, it's sad that Intel doesn't support you.

+++ Sources/Voxelotl/Renderer/Material.swift
@@ -1,7 +1,7 @@
 
 public struct Material {
-  public var ambient: Color<Float16>
-  public var diffuse: Color<Float16>
-  public var specular: Color<Float16>
+  public var ambient: Color<Float>
+  public var diffuse: Color<Float>
+  public var specular: Color<Float>
   public var gloss: Float
 }
